drop style constraint

+++ editor/src/components/text-editor/text-editor.tsx
@@ -333,6 +333,7 @@ const TextEditor = React.memo((props: TextEditorProps) => {
 
     currentElement.focus()
     savedContentRef.current = currentElement.textContent
+    const initialText = currentElement.textContent
 
     const elementCanvasFrame = MetadataUtils.getFrameOrZeroRectInCanvasCoords(
       elementPath,
@@ -357,11 +358,15 @@ const TextEditor = React.memo((props: TextEditorProps) => {
           if (elementState != null && savedContentRef.current !== content) {
             savedContentRef.current = content
             requestAnimationFrame(() => dispatch([getSaveAction(elementPath, content, textProp)]))
-
-            // remove dangling empty spans
-            if (content != null && content.replace(/^\n/, '').length === 0 && canDeleteWhenEmpty) {
-              requestAnimationFrame(() => dispatch([deleteView(elementPath)]))
-            }
+          }
+          // remove dangling empty spans
+          if (
+            content != null &&
+            initialText !== content &&
+            content.replace(/^\n/, '').length === 0 &&
+            canDeleteWhenEmpty
+          ) {
+            requestAnimationFrame(() => dispatch([deleteView(elementPath)]))
           }
         }
       }
@@ -608,21 +613,6 @@ function getSaveAction(
   return updateText(elementPath, escapeHTML(content, textProp), textProp)
 }
 
-const allowedStyleKeysForDeletion: string[] = [
-  'position',
-  'top',
-  'left',
-  'bottom',
-  'right',
-  'width',
-  'height',
-  'wordBreak',
-  'fontSize',
-  'fontWeight',
-  'fontFamily',
-  'font',
-]
-
 function canDeleteElementWhenEmpty(
   jsxMetadata: ElementInstanceMetadataMap,
   path: ElementPath,
@@ -638,14 +628,6 @@ function canDeleteElementWhenEmpty(
 
   const elementProps = allElementProps[toString(path)]
   if (elementProps == null) {
-    return false
-  }
-
-  // it must not have defined styling
-  if (
-    elementProps.style != null &&
-    Object.keys(elementProps.style).some((key) => !allowedStyleKeysForDeletion.includes(key))
-  ) {
     return false
   }
